Created a new project (SIMPLE TO DO APP)
Technologies used: HTML, CSS, JAVASCRIPT

Description:

A simple to do app using HTML, CSS and JavaScript.
Implemented use of localStorage. Also added comments
for each line of code for beginners aspiring to be a web developer.

Created the html file first and the Javascript 2nd
Will update the project for the styling.